added users and games endpoints and tests

+++ migrations/004.do.create_session_scores.sql
@@ -6,7 +6,7 @@ CREATE TABLE session_scores (
         REFERENCES games(id) ON DELETE CASCADE NOT NULL,
     uid INTEGER
         REFERENCES users(id) ON DELETE CASCADE,
-    score INTEGER NOT NULL,
+    score NUMERIC NOT NULL,
     name TEXT NOT NULL,
     winner BOOLEAN NOT NULL
 );
